requirements.txt
asgiref==3.6.0
autopep8==2.0.2
click==8.1.3
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
django-extra-fields==3.0.2
django-filter==22.1
django-grappelli==3.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.1.1
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
